Create Green Doc Template
A green doc is a preliminary design document that is used to gather feedback and input from stakeholders before the final design document is created. The green doc is typically less formal than the final design document and may not include all of the details that will be included in the final document. The purpose of the green doc is to get everyone on the same page about the project and to identify any potential problems or issues that need to be addressed before the final design is created.

Here are some of the benefits of using a green doc:

It can help to save time and money by identifying potential problems early on.
It can help to improve communication and collaboration between stakeholders.
It can help to ensure that the final design is well-thought-out and meets the needs of all stakeholders.
If you are working on a project that requires a design document, I recommend using a green doc to help you get started. It is a great way to gather feedback and input from stakeholders, identify potential problems, and improve communication and collaboration.

Here are some tips for writing a green doc:

Keep it short and to the point. The green doc is not meant to be a comprehensive document. It is simply a way to get everyone on the same page about the project.
Be clear and concise. The green doc should be easy to understand by everyone who will be reading it.
Use visuals. Visuals can help to communicate your ideas more effectively.
Get feedback. Once you have written a green doc, share it with stakeholders and get their feedback. This will help you to improve your document and make sure that it meets the needs of everyone involved.
___